rework save_model and  export function
